DCS-1474: ✨ add BaSM service
