+++ lib/generators/fromCourtWithPnc.js
@@ -1,80 +1,27 @@
 const {
-  random, data, profile, moves,
+  data,
 } = require('../data/models');
 
 const pncs = data.existingPNCs.map((i) => i.unknown)[Symbol.iterator]();
 
 class FromCourtWithPnc {
-  constructor(logger, basmApi) {
+  constructor(logger, basmApi, basmService) {
     this.logger = logger;
     this.api = basmApi;
-  }
-
-  async findPerson() {
-    const pnc = pncs.next().value;
-    const person = (
-      await this.api.getRaw(
-        `/api/people?filter[police_national_computer]=${pnc}`,
-      )
-    )[0];
-
-    const {
-      id, firstNames, lastName, dateOfBirth, gender,
-    } = person;
-
-    this.logger.info(
-      `Found person: ${firstNames} ${lastName} (${dateOfBirth}, ${gender})`,
-    );
-    return id;
-  }
-
-  async createProfile(personId) {
-    const { id } = await this.api.post(
-      `/api/people/${personId}/profiles`,
-      profile(),
-    );
-    this.logger.info(`Created profile: ${id}`);
-    return id;
-  }
-
-  async createMove(profileId) {
-    const move = moves.remand({
-      profile: profileId,
-      date: new Date(),
-      fromLocation: random(data.courts),
-      toLocation: data.toPrison.value,
-    });
-    const result = await this.api.post('/api/moves', move);
-    this.logger.info(`Created move: ${result?.id}`);
-    return result?.id;
-  }
-
-  async acceptMove(moveId) {
-    await this.api.post(`/api/moves/${moveId}/accept`, moves.accept());
-    this.logger.info(`Accepted move: ${moveId}`);
-  }
-
-  async startMove(moveId) {
-    await this.api.post(`/api/moves/${moveId}/start`, moves.start());
-    this.logger.info(`Started move: ${moveId}`);
-  }
-
-  async completeMove(moveId) {
-    await this.api.post(`/api/moves/${moveId}/complete`, moves.complete());
-    this.logger.info(`Completed move: ${moveId}`);
+    this.basmService = basmService;
   }
 
   async generate() {
-    const personId = await this.findPerson();
-    const profileId = await this.createProfile(personId);
-    const moveId = await this.createMove(profileId);
-    await this.acceptMove(moveId);
-    await this.startMove(moveId);
-    await this.completeMove(moveId);
+    const personId = await this.basmService.findPersonByPnc(pncs.next().value);
+    const profileId = await this.basmService.createProfile(personId);
+    const moveId = await this.basmService.createMoveRemand(profileId);
+    await this.basmService.acceptMove(moveId);
+    await this.basmService.startMove(moveId);
+    await this.basmService.completeMove(moveId);
   }
 }
-module.exports = async (logger, basmApi) => {
+module.exports = async (logger, basmApi, basmService) => {
   logger.info('Generating From Court Move with PNC');
-  const fromCourt = new FromCourtWithPnc(logger, basmApi);
+  const fromCourt = new FromCourtWithPnc(logger, basmApi, basmService);
   await fromCourt.generate();
 };
