+++ lib/generators/fromPoliceCustodySuite.js
@@ -1,73 +1,26 @@
 const {
-  random, data, moves, syncProfile,
+  random, data,
 } = require('../data/models');
 
 class FromPoliceCustodySuite {
-  constructor(logger, basmApi) {
+  constructor(logger, basmApi, basmService) {
     this.logger = logger;
     this.api = basmApi;
-  }
-
-  async findPerson() {
-    const prisonNumber = random(data.existingPrisonerNumbers);
-    const { id } = (
-      await this.api.getRaw(`/api/people?filter[prison_number]=${prisonNumber}`)
-    )[0];
-
-    this.logger.info(
-      `Found person with offender number: ${prisonNumber} and id: ${id})`,
-    );
-    return id;
-  }
-
-  async syncProfile(personId) {
-    const { id } = await this.api.post(
-      `/api/people/${personId}/profiles`,
-      syncProfile(),
-    );
-    this.logger.info(`Created profile: ${id}`);
-    return id;
-  }
-
-  async createMove(profileId) {
-    const move = moves.recall({
-      profile: profileId,
-      date: new Date(),
-      fromLocation: random(data.policeCustodaySuites),
-      toLocation: data.toPrison.value,
-    });
-
-    const result = await this.api.post('/api/moves', move);
-    this.logger.info(`Created move: ${result?.id}`);
-    return result?.id;
-  }
-
-  async acceptMove(moveId) {
-    await this.api.post(`/api/moves/${moveId}/accept`, moves.accept());
-    this.logger.info(`Accepted move: ${moveId}`);
-  }
-
-  async startMove(moveId) {
-    await this.api.post(`/api/moves/${moveId}/start`, moves.start());
-    this.logger.info(`Started move: ${moveId}`);
-  }
-
-  async completeMove(moveId) {
-    await this.api.post(`/api/moves/${moveId}/complete`, moves.complete());
-    this.logger.info(`Completed move: ${moveId}`);
+    this.basmService = basmService;
   }
 
   async generate() {
-    const personId = await this.findPerson();
-    const profileId = await this.syncProfile(personId);
-    const moveId = await this.createMove(profileId);
-    await this.acceptMove(moveId);
-    await this.startMove(moveId);
-    await this.completeMove(moveId);
+    // eslint-disable-next-line max-len
+    const personId = await this.basmService.findPersonByPrisonNumber(random(data.existingPrisonerNumbers));
+    const profileId = await this.basmService.syncProfile(personId);
+    const moveId = await this.basmService.createMoveRecall(profileId);
+    await this.basmService.acceptMove(moveId);
+    await this.basmService.startMove(moveId);
+    await this.basmService.completeMove(moveId);
   }
 }
-module.exports = async (logger, basmApi) => {
+module.exports = async (logger, basmApi, basmService) => {
   logger.info('Generating From Police Custody Suite Move');
-  const fromPoliceCustodySuite = new FromPoliceCustodySuite(logger, basmApi);
+  const fromPoliceCustodySuite = new FromPoliceCustodySuite(logger, basmApi, basmService);
   await fromPoliceCustodySuite.generate();
 };
