+++ lib/generators/index.js
@@ -3,20 +3,21 @@ const fromCourtWithPnc = require('./fromCourtWithPnc');
 const fromPoliceCustodySuite = require('./fromPoliceCustodySuite');
 
 module.exports = class Generator {
-  constructor(logger, api) {
+  constructor(logger, api, basmService) {
     this.api = api;
     this.logger = logger;
+    this.basmService = basmService;
   }
 
-  fromCourt() {
-    fromCourt(this.logger, this.api);
+  async fromCourt() {
+    await fromCourt(this.logger, this.api, this.basmService);
   }
 
-  fromCourtWithPnc() {
-    fromCourtWithPnc(this.logger, this.api);
+  async fromCourtWithPnc() {
+    await fromCourtWithPnc(this.logger, this.api, this.basmService);
   }
 
-  fromPoliceCustodySuite() {
-    fromPoliceCustodySuite(this.logger, this.api);
+  async fromPoliceCustodySuite() {
+    await fromPoliceCustodySuite(this.logger, this.api, this.basmService);
   }
 };
